pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["stefan dragicevic <stefan.dragicevic@vegait.rs>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^22.3"
python-dotenv = "^1.1.1"
requests = "^2.32.4"
pandas = "^2.3.1"
openai = "^1.98.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
